Move computePointPos to sketch.ts
